foomatic-ppdfile: subroutine for `foomatic.conf` parsing
The new subroutine parses the config file and returns
the parameters `onlyrecommended` and `listreadymadeppds`,
like the subroutine `cupslistppds` previously did.
The latter now simply calls the former to get those values.
This is in preparation of a ppd file mass export
functionality which will utilize the new subroutine
to decide which ppd files are to be generated.