foomatic-ppdfile: move ppd file rendering into own subroutine
The new subroutine `ppd_text_from_poid_and_driver` renders the
content of a ppd file, based on the printer id and driver name.
This previously happened directly in the `generateppd`
subroutine; now the latter subroutine just calls the former
to obtain the ppd file content and print it to stdout.

Most sanity tests (compatibility of the printer/driver pair)
are also moved into the new subroutine.
If the new subroutine identifies a problem, it emmits
a warning, but doesn't abort the program by itself.
Instead it just returns nothing
to indicate failure to the caller;
it is up to the caller to decide how to handle a failure.
The only caller (`generateppd`) does what
it previously did on a failure: It dies
(without a specific message as that got printed
already by `ppd_text_from_poid_and_driver`).

One exception: If a dummy driver name
("default" or "recommended") is provided
to `ppd_text_from_poid_and_driver`,
it dies hard by itself, since it assumes the caller
resolves these names into proper driver names.

This commit is in preparation of a ppd file mass export
functionality which will utilize the new subroutine
to generate lots of ppd files and to fail more
gracefully if one ppd file cannot be rendered.