foomatic-ppdfile: add `export` subcommand to get all ppd files
The commit at hand adds the `export` subcommand:
It will enumerate and generate all possible ppd files.
The `foomatic.conf` file permits to expand the generation to
ready-made ppd file, or to restrict it to recommended drivers.
The command line switch `-P` restricts processing to
ppd files where "make" or "model" match a given regex.
To speed up the generation, another command line switch `-N`
can be used to state the number of parallel processes;
this feature requires the MCE module to be installed
(otherwise, this module is not needed).

The `export` command will create a subdirectory for each
"make" and a subsubdirectory for each "model" in the current
working directory, and will place all ppd files there.
This might seem over-engineered, but it is usually much easier
to flatten such an over-engineered directory tree with a simple
shell script then to create the tree if it is really needed.

The new subcommand might help distributions to provide
ppd file packages based on the foomatic database.